requirements.txt
absl-py==2.1.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
idna==3.7
itsdangerous==2.2.0
jax==0.4.26
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
threadpoolctl==3.4.0
urllib3==2.2.1
Werkzeug==3.0.2
xgboost==2.0.3